+++ packages/components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.4](https://github.com/standardnotes/app/compare/@standardnotes/components-meta@2.5.3...@standardnotes/components-meta@2.5.4) (2022-06-29)
+
+**Note:** Version bump only for package @standardnotes/components-meta
+
+## [2.5.3](https://github.com/standardnotes/app/compare/@standardnotes/components-meta@2.5.2...@standardnotes/components-meta@2.5.3) (2022-06-29)
+
+**Note:** Version bump only for package @standardnotes/components-meta
+
 ## [2.5.2](https://github.com/standardnotes/app/compare/@standardnotes/components-meta@2.5.1...@standardnotes/components-meta@2.5.2) (2022-06-28)
 
 ### Bug Fixes