+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.26](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.2.25...@standardnotes/releases@1.2.26) (2022-06-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.2.25](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.2.24...@standardnotes/releases@1.2.25) (2022-06-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.2.24](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.2.23...@standardnotes/releases@1.2.24) (2022-06-29)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.2.23](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.2.22...@standardnotes/releases@1.2.23) (2022-06-28)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.2.22](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.2.21...@standardnotes/releases@1.2.22) (2022-06-28)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.2.21](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.2.20...@standardnotes/releases@1.2.21) (2022-06-28)
 
 **Note:** Version bump only for package @standardnotes/releases