+++ packages/web/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.27.5](https://github.com/standardnotes/app/compare/@standardnotes/web@3.27.4...@standardnotes/web@3.27.5) (2022-06-29)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.27.4](https://github.com/standardnotes/app/compare/@standardnotes/web@3.27.3...@standardnotes/web@3.27.4) (2022-06-29)
+
+### Bug Fixes
+
+* desktop backups directory change ([#1178](https://github.com/standardnotes/app/issues/1178)) ([b8efd82](https://github.com/standardnotes/app/commit/b8efd823ac7ab5997b92e74d7c99e8487b69c326))
+
+## [3.27.3](https://github.com/standardnotes/app/compare/@standardnotes/web@3.27.2...@standardnotes/web@3.27.3) (2022-06-28)
+
+### Bug Fixes
+
+* **mobile:** turn native textview into uncontrolled component ([#1176](https://github.com/standardnotes/app/issues/1176)) ([d198625](https://github.com/standardnotes/app/commit/d198625cfab1d3fb579b897b748699f021e26578))
+
+## [3.27.2](https://github.com/standardnotes/app/compare/@standardnotes/web@3.27.1...@standardnotes/web@3.27.2) (2022-06-28)
+
+### Bug Fixes
+
+* search bar ([#1175](https://github.com/standardnotes/app/issues/1175)) ([5e5cf61](https://github.com/standardnotes/app/commit/5e5cf618c5d11fb40f3dc280f7dda50e6eab7a72))
+
 ## [3.27.1](https://github.com/standardnotes/app/compare/@standardnotes/web@3.27.0...@standardnotes/web@3.27.1) (2022-06-28)
 
 ### Bug Fixes